Reassign MS training data in models_class1_pan
